# Fixed #28459 -- Improved performance of ValuesListIterable.